fix ublox UBX-NAV-HPPOSLLH decoder

+++ sunray/src/ublox/ublox.cpp
@@ -56,7 +56,7 @@ void UBLOX::begin(){
   CONSOLE.print("test relPosN=");
   CONSOLE.println(relPosN);
   
-  // relPosE... PAYLOAD: ofs=12size=4    95,F9,FF,FF,     relPosE: 42949656.00  
+  // relPosE... PAYLOAD: ofs=12size=4    95,F9,FF,FF,     relPosE: 42949656.00   (wrong)
   payload[12] = 0x95;
   payload[13] = 0xF9;
   payload[14] = 0xFF;
@@ -442,9 +442,9 @@ void UBLOX::dispatchMessage() {
           case 0x14: 
             { // UBX-NAV-HPPOSLLH
               iTOW = (unsigned long)this->unpack_int32(4);
-              lon = (1e-7  * (this->unpack_int32(8)   +  (this->unpack_int8(24) * 1e-2)));
-              lat = (1e-7  * (this->unpack_int32(12)  +  (this->unpack_int8(25) * 1e-2)));
-              height = (1e-3 * (this->unpack_int32(16) +  (this->unpack_int8(26) * 1e-2))); // HAE (WGS84 height)
+              lon = 1e-7  * (    ((float)((int)this->unpack_int32(8)))   +  ((float)((int)this->unpack_int8(24))) * 1e-2    );
+              lat = 1e-7  *  (   ((float)((int)this->unpack_int32(12)))   +  ((float)((int)this->unpack_int8(25))) * 1e-2   );
+              height = 1e-3 * (  ((float)((int)this->unpack_int32(16))) +  ((float)((int)this->unpack_int8(26))) * 1e-2    ) ; // HAE (WGS84 height)
               //height = (1e-3 * (this->unpack_int32(20) +  (this->unpack_int8(27) * 1e-2))); // MSL height
               hAccuracy = ((double)((unsigned long)this->unpack_int32(28))) * 0.1 / 1000.0;
               vAccuracy = ((double)((unsigned long)this->unpack_int32(32))) * 0.1 / 1000.0;
